requirements.txt
anyio==3.6.2
click==8.1.3
distlib==0.3.6
et-xmlfile==1.1.0
fastapi==0.95.1
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
numpy==1.24.2
openpyxl==3.1.2
pkg_resources==0.0.0
protobuf==3.20.3
pydantic==1.10.7
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
unicorn==2.0.1.post1
uvicorn==0.21.1
